P14a Batch B: Timestamp cascade across 29 main-source files
Full-layer S11 propagation of BlockHeader.unixTimestamp: Long -> Timestamp.
.toLong only at actual wire/storage boundaries per the S11 protocol.

Consensus / DAA:
- DifficultyCalculator, EthashDifficultyCalculator, TargetTimeDifficultyCalculator:
  blockTimestamp: Long -> Timestamp; Timestamp-Timestamp=Long delta via extension
- BlockGeneratorSkeleton: wraps BlockTimestampProvider.getEpochSecond at call site
- NoOmmersBlockGenerator, PoWBlockGenerator: prepareHeader signature updated
- StdSignedTransactionValidator: private blockHeaderTimestamp params updated

Fork dispatch:
- BlockchainConfig: all six isXTimestamp(Long) -> isXTimestamp(Timestamp)
- BlockchainConfigForEvm: isPragueTimestamp/isOsakaTimestamp/isBpo1/isBpo2 updated
- EvmConfig.forBlock: timestamp param updated

Engine API (ETH wire boundary):
- EngineApiService: attrs.timestamp/payload.timestamp wrapped with Timestamp(...)
  at assignment; .toLong when comparing against wire Long
- EngineApiController: payload.timestamp and attrTimestamp (Option[Long]) wrapped

JSON-RPC / GraphQL (wire-out boundary):
- BlockResponse, TransactionReceiptResponse, TransactionResponse,
  FilterManager, EthSimulateService, EthSimulateJsonMethodsImplicits,
  GraphQLSchema: BigInt/Long wire outputs use .toLong
- EthTxService, BranchResolution: getOrElse(Timestamp.Zero)
- EthTxService: getOrElse(0L) -> getOrElse(Timestamp.Zero)

EVM:
- OpCode.TIMESTAMP: UInt256(s.env.blockHeader.unixTimestamp.toLong)
- BlockExecution: UInt256/BigInt EIP-4788 ring-buffer access uses .toLong

Networking:
- EthNodeStatus68ExchangeState, EthNodeStatus69ExchangeState: storedTimestamp
  typed as Timestamp; ForkId.create receives .toLong (wire Long param)

Storage / Genesis:
- GenesisDataLoader: genesisTimestamp promoted to Timestamp; unixTimestamp
  constructor arg already Timestamp(...)

MCP / History:
- McpTools: Instant.ofEpochSecond(.toLong)
- TransactionHistoryService: lazy timestamp: Long = header.unixTimestamp.toLong

+++ src/main/scala/com/chipprbots/ethereum/jsonrpc/BlockResponse.scala
@@ -125,7 +125,7 @@ object BlockResponse:
       size = Block.size(block),
       gasLimit = block.header.gasLimit.value,
       gasUsed = block.header.gasUsed.value,
-      timestamp = block.header.unixTimestamp,
+      timestamp = BigInt(block.header.unixTimestamp.toLong),
       mixHash = block.header.mixHash.value,
       transactions = transactions,
       uncles = block.body.uncleNodesList.map(_.hash.value),
